Abstracted out some stubs to reduce code duplication
