+++ src/test/java/seedu/address/logic/commands/ExportCommandTest.java
@@ -8,30 +8,24 @@ import java.io.IOException;
 import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.NoSuchFileException;
 import java.util.List;
-import java.util.function.Predicate;
 
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.converters.PersonConverter;
-import seedu.address.logic.converters.exceptions.PersonDecodingException;
 import seedu.address.logic.converters.exceptions.PersonEncodingException;
 import seedu.address.logic.converters.fileformats.AdaptedPerson;
-import seedu.address.logic.converters.fileformats.SupportedFile;
-import seedu.address.logic.converters.fileformats.SupportedFileFormat;
 import seedu.address.logic.converters.fileformats.csv.CsvAdaptedPerson;
-import seedu.address.model.Model;
-import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.event.Event;
 import seedu.address.model.person.Person;
-import seedu.address.model.tag.Tag;
 import seedu.address.testutil.TypicalPersons;
+import seedu.address.testutil.stubs.CsvConverterStub;
+import seedu.address.testutil.stubs.CsvFileStub;
+import seedu.address.testutil.stubs.ModelStubContainingTypicalPersons;
+import seedu.address.testutil.stubs.PersonConverterStub;
+import seedu.address.testutil.stubs.SupportedFileStub;
 
 public class ExportCommandTest {
 
@@ -101,47 +95,7 @@ public class ExportCommandTest {
                 TypicalPersons.NUM_PERSONS, VALID_CSV_FILENAME), commandResult.feedbackToUser);
     }
 
-    /**
-     * A default SupportedFile stub that have all of the methods failing.
-     */
-    private class SupportedFileStub implements SupportedFile {
-        @Override
-        public List<AdaptedPerson> readAdaptedPersons() throws IOException {
-            throw new AssertionError("This method should not be called.");
-        }
 
-        @Override
-        public void writeAdaptedPersons(List<AdaptedPerson> adaptedPersons) throws IOException {
-            throw new AssertionError("This method should not be called.");
-
-        }
-
-        @Override
-        public SupportedFileFormat getSupportedFileFormat() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public String getFileName() {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
-     * A CsvFile stub, an implementation of SupportedFile in the CSV format
-     */
-    private class CsvFileStub extends SupportedFileStub {
-        @Override
-        public SupportedFileFormat getSupportedFileFormat() {
-            return SupportedFileFormat.CSV;
-        }
-
-        @Override
-        public String getFileName() {
-            return VALID_CSV_FILENAME;
-        }
-
-    }
 
     /**
      * A CsvFile stub that have an already existing file
@@ -182,184 +136,31 @@ public class ExportCommandTest {
         public void writeAdaptedPersons(List<AdaptedPerson> adaptedPersons) throws IOException {
             // called by {@code ExportCommand#execute()}
         }
-    }
-
-
-    /**
-     * A default PersonConverter stub that have all of the methods failing.
-     */
-    private class PersonConverterStub implements PersonConverter {
-        @Override
-        public AdaptedPerson encodePerson(Person person) throws PersonEncodingException {
-            throw new AssertionError("This method should not be called.");
-        }
 
         @Override
-        public Person decodePerson(AdaptedPerson person) throws PersonDecodingException {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public SupportedFileFormat getSupportedFileFormat() {
-            throw new AssertionError("This method should not be called.");
+        public String getFileName() {
+            return VALID_CSV_FILENAME;
         }
     }
 
-    /**
-     * A CsvConverter stub, an implementation of PersonConverter in the CSV format
-     */
-    private class CsvConverterStub extends PersonConverterStub {
-        @Override
-        public SupportedFileFormat getSupportedFileFormat() {
-            return SupportedFileFormat.CSV;
-        }
-
-    }
 
     /**
      * A CsvConverter stub that always fails to encode
      */
-    private class CsvConverterAlwaysFailingConversion extends PersonConverterStub {
+    private class CsvConverterAlwaysFailingConversion extends CsvConverterStub {
         @Override
         public AdaptedPerson encodePerson(Person person) throws PersonEncodingException {
             throw new PersonEncodingException("Person fails to encode");
-        }
-
-        @Override
-        public SupportedFileFormat getSupportedFileFormat() {
-            return SupportedFileFormat.CSV;
         }
     }
 
     /**
      * A CsvConverter stub that always successfully encodes.
      */
-    private class CsvConverterAlwaysSuccessfulConversion extends PersonConverterStub {
+    private class CsvConverterAlwaysSuccessfulConversion extends CsvConverterStub {
         @Override
         public AdaptedPerson encodePerson(Person person) throws PersonEncodingException {
             return new CsvAdaptedPerson("Encoded person");
-        }
-
-        @Override
-        public SupportedFileFormat getSupportedFileFormat() {
-            return SupportedFileFormat.CSV;
-        }
-    }
-
-
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void addPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addEvent(Event event) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void resetData(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasPerson(Person person) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasEvent() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deletePerson(Person target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteEvent() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updatePerson(Person target, Person editedPerson) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateEvent(Event editedEvent) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteTag(Tag tag) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addTag(Tag tag) {
-            throw new AssertionError("This method should not be called");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Event getEventDetails() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean canUndoAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean canRedoAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void undoAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void redoAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void commitAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
-     * A Model stub that contains TypicalPersons
-     */
-    private class ModelStubContainingTypicalPersons extends ModelStub {
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            return FXCollections.observableList(TypicalPersons.getTypicalPersons());
         }
     }
 }
